packages/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared infrastructure drivers and cross-cutting contracts"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.70.0",
    "sqlalchemy>=2.0.44",
    "asyncpg>=0.31.0",
    "psycopg[binary]>=3.2.12",
    "redis>=6.4.0",
    "pydantic-settings>=2.8.0",
    "pinecone>=8.1.0",
    "qdrant-client>=1.12.0",
    "minio>=7.2.18",
    "pika>=1.3.2",
    "tiktoken>=0.12.0",
    "numpy>=1.26.0",
    "bcrypt>=5.0.0",
    "PyJWT>=2.10.1",
    "fastapi>=0.135.1",
    "structlog>=24.0.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]
